HAIA/requirements.txt
torch==1.8.1
scikit_image==0.18.1
matplotlib==3.4.1
numpy==1.19.2
gym==0.18.0
tqdm==4.60.0
torchvision==0.9.1
ipython==7.23.1
Pillow==8.2.0
skimage==0.0
